update tests and fix component per file
